Moved max block size value into constants.
