+++ constants/BlockConstants.ts
@@ -1,5 +1,7 @@
+export const MAX_BLOCK_SIZE = 524288
+
 export enum BlockMetricGranularity {
   Day = 'day',
   Lifetime = 'lifetime',
-  Total = 'total'
+  Total = 'total',
 }
